Release of 4.03.

+++ src/main/java/de/mossgrabers/framework/controller/DefaultValueChanger.java
@@ -12,7 +12,7 @@ package de.mossgrabers.framework.controller;
 public class DefaultValueChanger implements IValueChanger
 {
     private final int    upperBound;
-    private final int    fractionValue;
+    private final double fractionValue;
     private final double slowFractionValue;
     private boolean      isSlow;
 
@@ -24,7 +24,7 @@ public class DefaultValueChanger implements IValueChanger
      * @param fractionValue Amount by which values are incremented / decremented
      * @param slowFractionValue Amount by which values are slowly incremented / decremented
      */
-    public DefaultValueChanger (final int upperBound, final int fractionValue, final double slowFractionValue)
+    public DefaultValueChanger (final int upperBound, final double fractionValue, final double slowFractionValue)
     {
         this.upperBound = upperBound;
         this.fractionValue = fractionValue;
@@ -42,7 +42,7 @@ public class DefaultValueChanger implements IValueChanger
 
     /** {@inheritDoc} */
     @Override
-    public int getFractionValue ()
+    public double getFractionValue ()
     {
         return this.fractionValue;
     }
